+++ flutter_sound/example/lib/volume_control/volume_control.dart
@@ -30,18 +30,21 @@ import 'package:flutter_sound/flutter_sound.dart';
  *
  */
 
-final _exampleAudioFilePathMP3_1 =
-    'https://flutter-sound.canardoux.xyz/web_example/assets/extract/05.mp3';
-final _exampleAudioFilePathMP3_2 =
-    'https://flutter-sound.canardoux.xyz/web_example/assets/extract/13.wav';
+const _exampleAudioFilePathMP3_1 =
+    'https://flutter-sound.canardoux.xyz/extract/05.mp3';
+
+const _exampleAudioFilePathWAV_2 =
+    'https://flutter-sound.canardoux.xyz/extract/13.wav';
 
 ///
 typedef Fn = void Function();
 
 /// Example app.
 class VolumeControl extends StatefulWidget {
+  const VolumeControl({super.key});
+
   @override
-  _VolumeControlState createState() => _VolumeControlState();
+  State<VolumeControl> createState() => _VolumeControlState();
 }
 
 class _VolumeControlState extends State<VolumeControl> {
@@ -141,11 +144,11 @@ class _VolumeControlState extends State<VolumeControl> {
       return Container(
         margin: const EdgeInsets.all(3),
         padding: const EdgeInsets.all(3),
-        height: 240,
+        height: 250,
         width: double.infinity,
         alignment: Alignment.center,
         decoration: BoxDecoration(
-          color: Color(0xFFFAF0E6),
+          color: const Color(0xFFFAF0E6),
           border: Border.all(
             color: Colors.indigo,
             width: 3,
@@ -159,14 +162,14 @@ class _VolumeControlState extends State<VolumeControl> {
               //disabledColor: Colors.grey,
               child: Text(_mPlayer1.isPlaying ? 'Stop' : 'Play'),
             ),
-            SizedBox(
+            const SizedBox(
               width: 20,
             ),
             Text(_mPlayer1.isPlaying
                 ? 'Playback #1 in progress'
                 : 'Player #1 is stopped'),
           ]),
-          Text('Volume:'),
+          const Text('Volume:'),
           Slider(
               value: _mVolume1,
               min: 0.0,
@@ -175,19 +178,19 @@ class _VolumeControlState extends State<VolumeControl> {
               divisions: 100),
           Row(children: [
             ElevatedButton(
-              onPressed: getPlaybackFn(_mPlayer2, _exampleAudioFilePathMP3_2),
+              onPressed: getPlaybackFn(_mPlayer2, _exampleAudioFilePathWAV_2),
               //color: Colors.white,
               //disabledColor: Colors.grey,
               child: Text(_mPlayer2.isPlaying ? 'Stop' : 'Play'),
             ),
-            SizedBox(
+            const SizedBox(
               width: 20,
             ),
             Text(_mPlayer2.isPlaying
                 ? 'Playback #2 in progress'
                 : 'Player #2 is stopped'),
           ]),
-          Text('Volume:'),
+          const Text('Volume:'),
           Slider(
               value: _mVolume2,
               min: 0.0,