+++ flutter_sound_platform_interface/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 9.2.13
+## 9.4.6
 
-- Please [look to this](https://dooboolab.github.io/flutter_sound/book/CHANGELOG.html)
+- Please [look to this](https://flutter-sound.canardoux.xyz/changelog.html)